add metadata.yaml to slice outputs (#965)
### Description

This adds a metadata.yaml file to slicez_pltxxxxx folders. This is
helpful because when yt loads that folder (it can, very sweet), yt will
recognize it as a Quokka simulation output and do all the customization
for Quokka.

Note that this yt customization for Quokka is implemented and tested in
[my yt fork](https://github.com/chongchonghe/yt) and we are in the
process of publishing it before the next yt release. Demo of using the
new YT frontend is here:
https://github.com/Rongjun-ANU/README-of-yt-frontend-for-QUOKKA/blob/main/README.ipynb

Other changes: 
- this also adds `tests/slice*` to .gitignore

### Related issues
None.

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues (if applicable; see
above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [ ] I have added tests for any new physics that this PR adds to the
code.
- [ ] *(For quokka-astro org members)* I have manually triggered the GPU
tests with the magic comment `/azp run`.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>