binary orbit test on multilevel grids (#983)
### Description

Previously the binary_orbit test runs on uniform grid. This PR add a new
test with the whole domain refined to `lev=1`. The orbit separation is
checked against the expected value and the error is within 6%. Note that
the orbit radius is resolved by only 4 cells. In the uniform-grid test,
the orbit radius is resolved by 2 (!!!) cells and the the error is
confined in 16%.

More changes: 
- I also updated `computeAfterTimestep` in `binary_orbit.cpp` to copy
all particles from the finest level to the IO processor and validate the
answer on the go. Previously `computeAfterTimestep` only works when AMR
is disabled (`lev_max = 0`).

### Related issues
Are there any GitHub issues that are fixed by this pull request? Add a
link to them here.

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues (if applicable; see
above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [x] I have added tests for any new physics that this PR adds to the
code.
- [x] *(For quokka-astro org members)* I have manually triggered the GPU
tests with the magic comment `/azp run`.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>