Define a virtual function to tell if a particle type belongs to the Star particle class (#987)
### Description
We have `PhysicsParticleDescriptor` for most particle types and
`StarParticleDescriptor` for star-like particle types, and some
operations like `depositSN` only exists in the latter. However, the code
does not know which of these two classes a `descriptor` belongs to. In
this PR, I define a virtual function `isStarParticle()` that returns
whether or not a particle type belongs to `StarParticleDescriptor` so
that we can call `depositSN` on it. This makes `depositSN` in
`PhysicsParticleRegister` very clean.

This design can generalize to more particle classes , in which case I
will change `isStarParticle()` to `getParticleClass` which returns the
particle class: base_particles, star_particles, DM_particles, etc.

### Related issues
Are there any GitHub issues that are fixed by this pull request? Add a
link to them here.

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues (if applicable; see
above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [x] I have added tests for any new physics that this PR adds to the
code.
- [x] *(For quokka-astro org members)* I have manually triggered the GPU
tests with the magic comment `/azp run`.