Add tests of SNR deposition on multiple levels (#964)
### Description

In this PR, all particle operations (creation, destruction and SNR
deposition) are fully tested on multiple levels. Assuming particles
always live in the finest level, SNR deposition is accurate to machine
accuracy even if they are near the coarse-fine boundary. Particles in
coarser grids near coarse-fine boundary might work just fine, although I
haven't tested it because the gravity solver doesn't work in this
situation.

Now we have two problems to test particle operations:

1. `Gravity3D`. This problem tests particle creation from fluid cells,
stellar evolution (from SNProgenitor to SNRemnant), supernova
deposition, and stellar destruction with fully refined grids (lev=1 in
all domain). In the end of the simulation, we check (1) the number of
Test particles, (2) the SNR mass, and (3) the orbit of the CIC particles
(binary star in circular orbit), which is checked agains analytic
solution. The Test particles are located at the corners of MultiFab
boxes where eight boxes intersect. SNR mass is accurate to machine
accuracy.
2. `Gravity3DAMR`. This is the same test problem but with half of the
domain refined to lev=1 and the other half in lev=0. The CIC particle is
also removed due to the failure of the gravity solver when a particle
crosses a coarse-fine boundary. The Test particles are located at the
intersections between MultiFab boxes, and also at the coarse-fine
boundary. SNR deposition is accurate to machine accuracy at the step
right after deposition, but the relative error increases to ~1e-6
several steps after. Is mass a conserved quantity when AMR is turned on?
2. Rewrite `getParticleData()` (now named `getParticleDataAtLevel()`) to
make it work on any AMR level.
4. Removed `getParticlePositions()` since `getParticleDataAtLevel()`
provides the same (and more) functionality.

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues (if applicable; see
above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [x] I have added tests for any new physics that this PR adds to the
code.
- [x] *(For quokka-astro org members)* I have manually triggered the GPU
tests with the magic comment `/azp run`.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>