Remove default CUDA GPU architectures
This removes the default settings for CUDA GPU architectures. (We have never set a default for AMD GPUs.)

This is done to reduce the compile time, since the default was set to 2 GPU architectures, which made the build time almost twice as large as it would be compared to setting only one GPU arch. This makes the GPU tests on avatargpu take significantly longer.

Users will need to specify which CUDA GPU architecture they want to build for explicitly using either `-DAMReX_CUDA_ARCH` or the `AMREX_CUDA_ARCH` environment variable.