A new SN feedback model (#995)
### Description

In this PR I implement a new SNe feedback model that smoothly
transitions between thermal feedback, kinetic feedback, and momentum
feedback, depending on how well the cooling/shell formation phase of the
SNR evolution is resolved. An innovation of this scheme is that we do a
two stage deposition scheme to account for SNR overlapping and get rid
of depedence on the order that SNe events occur.

I have tested three models:
1. Model 0: Pure thermal deposition (for comparison only, not for
production)
2. Model 1: Thermal energy deposition in the resolved regime, thermal +
kinetic in the intermediate regime, and terminal momentum in the
unresolved regime.
3. Model 2: Kinetic energy deposition in the resolved regime, thermal +
kinetic in the intermediate regime, and terminal momentum in the
unresolved regime.

At a SN event, we distribute mass, momentum, and energy uniformly among
the cells that intercect with s sphere of radius `3 dx` centered at the
center of the cell that the SN is located in. No background smoothing is
conducted before deposition.

The results looks great! See figure below. Both model 1 and 2 produces
'correct' and almost identical results. The only difference is the
maximum temperature (defined as the maximum temperature in any cell
throughout the simulation) in model 2 is slightly lower than that in
model 1. In model 2, right after SN explosion, the gas is cool but moves
fast, and the shock wave immediately heat the gas to comparable
temperature to that in the thermal model. After some experiments, I am
convinced that the thermal model and kinetic model (in the resolved
regime) has no difference in accuracy or performance related to
timestep. This is because, imagine a star explodes into vacuum, q simple
calculation will show that the sound speed of the hot SNR in the pure
thermal model is about $1/\sqrt{3}$ times the bulk velocity of the cool
SNR in the pure kinetic model (microscopic thermal motion has three
degrees of freedom while kinetic motion has one). The thermal model
turns out to have slightly longer timestep overall. In the TIGRESS model
they advocate thermal model because they have over cooling issue, but we
don't have this issue at all (thanks to Ben's awesome Grackle cooling
module which has a robust ODE solver).


![compare](https://github.com/user-attachments/assets/23895e3d-e00a-46a5-af30-a62d70509a02)

> Reproducing Figure 6 of Kim & Ostriker 2017. The spacial resolution
and SN ejecta mass/energy/momentum in our simulations are the same as
theirs.

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues (if applicable; see
above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [x] I have added tests for any new physics that this PR adds to the
code.
- [ ] *(For quokka-astro org members)* I have manually triggered the GPU
tests with the magic comment `/azp run`.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>