Allow to force some particles to always live in finest cells (#1009)
### Description
We want to force some particles to always live on the finest cells.
Those particles could be sink particles that need to accrete, or
`SNProgenitor` particles that can go SN explosion at any time. This is
also important for getting around the issue with the gravity solver on
refinment boundaries.

The user can turn on manually turn on this feature for a specific
particle type by running the following line of code in `problem_main()`

```cpp
sim.particleRegister_.getParticleDescriptor(quokka::ParticleType::Test)->setForceFinestLevel(true);
```

Here is a demonstration with the `ParticleCreation` test. It seems that
by tagging a single cell at the particle location, the code actually
refines at least 4+ cells around it. This happens to be the desired
behaviour. Lovely.


![plt00004_Slice_z_density](https://github.com/user-attachments/assets/a31caa63-51da-45ea-b403-a830e1576c57)


Note: This branch is set to merge to #1005 because it's based on that
branch. This is just for showing the actual changes made in this PR.
Once we merge #1005 , I'll update this PR to merge to dev branch.

### Related issues
None. 

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues see (see above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [x] I have added tests for any new physics that this PR adds to the
code.
- [ ] I have tested this PR on my local computer and all tests pass.
- [x] I have manually triggered the GPU tests with the magic comment
`/azp run`.
- [x] I have requested a reviewer for this PR.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>