Stochastic Stellar Population (#966)
### Description
This branch includes a specialisation for the particles. We check cells
for Jeans stability condition and if it is Jeans unstable, we add a
population of particles. The population comprises one particle
representing low mass star and a random number of high mass particles.
The number of high mass particles is determined using the Chabrier IMF,
which is log normal for M<Msun and a power law with a slope of -2.35
above that. It is assumed that the IMF covers a mass range 0.08-120
Msun. The fate and lifetime of the high mass particles are interpolated
using tables from SLUG.

Once added, the high mass stars gain a momentum weighted over
neighbouring particles while the low mass particle gains an opposite
momentum. At the end of particle updates, the density, total energy and
the internal energy density of the gas cell.

### Related issues
N/A

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [ ] I have added a link to any related issues (if applicable; see
above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [ ] I have added tests for any new physics that this PR adds to the
code.
- [x] *(For quokka-astro org members)* I have manually triggered the GPU
tests with the magic comment `/azp run`.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Ben Wibking <ben@wibking.com>
Co-authored-by: ChongChong He <chongchong.he@anu.edu.au>