Add a simple script to run clang-tidy on modified files (#1056)
### Description

This will run `clang-tidy` on all the files that have been modified:
```bash
./scripts/tidy.sh ./build
```
where `./build` is the directory to the build folder.

This will run `clang-tidy` on all the files that have been modified in
the last commit:
```bash
./scripts/tidy.sh ./build previous
```

This will run `clang-tidy` on all the files that have been modified with
respect to the remote branch:
```bash
./scripts/tidy.sh ./build origin
```

This will run `clang-tidy` on all the files that have been modified with
respect to the local development branch:
```bash
./scripts/tidy.sh ./build dev
```

Also update the documentation on how to use tidy. The python script Ben
originally recommended is too complicated and I couldn't make it work
the way I wanted. The new bash script is a lot simpler and easy to use.

### Related issues
Are there any GitHub issues that are fixed by this pull request? Add a
link to them here.

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [ ] I have added a description (see above).
- [ ] I have added a link to any related issues (if applicable; see
above).
- [ ] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [ ] I have added tests for any new physics that this PR adds to the
code.
- [ ] *(For quokka-astro org members)* I have manually triggered the GPU
tests with the magic comment `/azp run`.