Frontier-like docker container (#1088)
### Description
This changes the ROCm container to be as close as possible to the
environment on Frontier and Setonix:
* **OpenSUSE Leap 15.6** (Frontier uses the commerical SLES 15.6).
* **MPICH 3.4.3** (from which Cray MPI is derived and which which it is
binary compatible). This lets us compile against MPICH in the container
and then, when running on Frontier, bind-mount the Cray MPI libraries
and use them at runtime.
* **ROCm 6.4.0** Should be the same everywhere, in principle, but OLCF
rebuilds this from source. We can just use the version in the container.

Things that will not be identical:
* Cray MPI vs MPICH: Cray MPI has optimizations and bugs that aren't in
open-source MPICH. For testing purposes only, we can run with
open-source MPICH across multiple systems (this will be slow and should
never be done for production sims). Usually, the bugs we've seen aren't
in MPI itself, but in the network layer.
* Network: Setonix and Frontier have HPE Slingshot with different driver
and firmware versions installed. It's buggy and behaves differently
across versions, and we can't do anything about it.
* GPU driver: different systems have different driver versions
installed. We can't do anything about this.
 
### Related issues
https://github.com/quokka-astro/quokka/issues/1087

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues (if applicable; see
above).
- [ ] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [ ] I have added tests for any new physics that this PR adds to the
code.
- [ ] *(For quokka-astro org members)* I have manually triggered the GPU
tests with the magic comment `/azp run`.