Fix bugs in SF and SN prescriptions (#1137)
### Description

This PR fixes one fatal bug in the SF prescription and one (potentially)
fatal bug in the SN prescription.

Bug fix in SF: The turbulent velocity in a star-forming cell could be
zero, causing a call of RandomNormal with sigma = 0, which is illegal. I
set a floor of `sigma_sq = cs * cs`, which reflects that star-forming
regions typically have turbulent and thermal energies in equipartition.

Bug fix in SN: During the limiter step, I added a special case for `a =
0` in the solution of `a x^2 + b x + c = 0`. I don't see it happen in
the test problem, and it could potentially happen somewhere.

Also updated the SF test problem:
- Add more checks at step 1: check the total mass in high-mass stars,
the number of high-mass stars, and the number of low-mass stars. Fail
the test if those numbers are not within tolerance.
- Let the problem run to step 10 and check mass conservation in the end.

### Related issues
None. 

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues (if applicable; see
above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [ ] I have added tests for any new physics that this PR adds to the
code.
- [x] *(For quokka-astro org members)* I have manually triggered the GPU
tests with the magic comment `/azp run`.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>