Extrema-preserving PPM (#1181)
### Description
This implements the 5th-order, extrema-preserving, hybrid PPM-WENO
reconstruction method of [Rider, Greenough, and Kamm
(2007)](https://ui.adsabs.harvard.edu/abs/2007JCoPh.225.1827R/abstract).

The number of ghost zones required is the same as classic PPM.

The default reconstruction method is unchanged (classic PPM). The
advection tests and `RadhydroShockCGS` are modified to use this method
by default, so this new method is tested in the test suite. (While
testing this, I also discovered the RadhydroShock tests were slightly
too under-resolved to get the Zeldovich spike right, so those are all
increased to 256 cells.)

Caveats:
* Whether this method perfectly maintains an isolated contact wave is
compiler-dependent. There appears to be sensitivity to how the
floating-point operations are grouped. This causes the `HydroContact`
test to fail for one CI configuration when xPPM is enabled for this
problem.
* The `RadhydroShell` problem crashes with this method. It might need to
be run at a lower CFL number.
* Note that only the dimension-by-dimension reconstruction is 5th order:
the overall order of accuracy in space in still 2nd order, since the
conservative-to-primitive conversion assumes that cell averages and cell
centers are identical (also the face center and face average fluxes are
assumed to be identical).

### Related issues
N/A

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues (if applicable; see
above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [ ] I have added tests for any new physics that this PR adds to the
code.
- [x] *(For quokka-astro org members)* I have manually triggered the GPU
tests with the magic comment `/azp run`.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>