Set star particle velocities to include runaways (#1236)
### Description
This PR modifies the way velocities of high mass stars are estimated.
Velocities are drawn from a power law with a slope of -1.8. The slope of
the distribution is tested in the ParticleSF test problem.

### Related issues
None.

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x ] I have added a description (see above).
- [ ] I have added a link to any related issues (if applicable; see
above).
- [x ] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [ x] I have added tests for any new physics that this PR adds to the
code.
- [ ] *(For quokka-astro org members)* I have manually triggered the GPU
tests with the magic comment `/azp run`.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>