update exit behavior when max_walltime is set (#1274)
### Description
Exiting at 90% of `max_walltime` is not very efficient when the job
walltime is large. Instead, we stop the simulation 5 minutes before
`max_walltime` (unless the walltime is less than 50 minutes, in which
case we fall back to stopping at 90% of the limit).

### Related issues
N/A

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues (if applicable; see
above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [ ] I have added tests for any new physics that this PR adds to the
code.
- [ ] *(For quokka-astro org members)* I have manually triggered the GPU
tests with the magic comment `/azp run`.