Making ParticleToMesh reproducible by rounding off rhs at 12th digit.  (#1323)
### Description

This completely solves the particle reproducibility issue on a single
level (multilevel yet to be tested).

The solution is simple. Assuming there is no more than an order of a few
particles in a cell, we can probably estimate the upper bound on the
roundoff error caused by float-point addition. From my tests, it's
always at or below 1e-13. We simply roundoff the `rhs_buffer` to the
12th digit before adding it to the real `rhs`. Since this is done after
`SumBoundary`, a deterministic `SumBoundary` is not required.

The loss of accuracy on `rhs_buffer` doesn't matter, because
particle-mesh interaction always have huge systematics. Particularly,
the CIC algorithm with cubic kernel is very crude and I speculate it's
systematics is at least 1% (Ben can confirm this). Other particle-mesh
interaction recipes, like sink accretion and SN, has a systematic of 1%
to 20%.

Tested with the spherical_collapse problem the same way I did in #1299.
The PR gives reproducible results with 100000 particles in a 32^3 box
with 4^3 fabs, i.e. ~3 particles per cell.

### Related issues

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [ ] I have added a description (see above).
- [ ] I have added a link to any related issues (if applicable; see
above).
- [ ] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [ ] I have added tests for any new physics that this PR adds to the
code.
- [ ] *(For quokka-astro org members)* I have manually triggered the GPU
tests with the magic comment `/azp run`.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>