Use deterministic AMR flux register (#1342)
### Description
`amrex::FluxRegister` is deterministic, while our previous
`amrex::YAFluxRegister` was not (due to using atomics). We need this to
be deterministic for AMR simulations to be reproducible.

Using the deterministic `amrex::FluxRegister` requires passing it a
MultiFab of fluxes, so this increases the memory usage of the radiation
update, which did not previously have such a MultiFab.

Depends on https://github.com/quokka-astro/quokka/pull/1343.

### Related issues
Are there any GitHub issues that are fixed by this pull request? Add a
link to them here.

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues (if applicable; see
above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [ ] I have added tests for any new physics that this PR adds to the
code.
- [ ] *(For quokka-astro org members)* I have manually triggered the GPU
tests with the magic comment `/azp run`.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>