trigger outputs when sentinel files are created in working dir (#1398)
### Description
This adds a check for the file `outputNow` in the simulation working
directory after each coarse step. If it exists, a plotfile is written
and then `outputNow` is deleted. An analogous feature is added for
checkpoints with the `checkpointNow` file.

The output code for plotfiles and checkpoints is modified to ensure that
only one checkpoint or plotfile is written per coarse step. Previously,
it was possible that multiple outputs of the same type were written if
both the step interval and simulation time interval were specified. Now
each output trigger checks whether an output of the same type has
already been triggered for the current step (for the step interval
trigger, the simulation time interval trigger, and the sentinel file
trigger).

This is inspired by the analogous Enzo feature.

### Related issues
N/A

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues (if applicable; see
above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [ ] I have added tests for any new physics that this PR adds to the
code.
- [ ] *(For quokka-astro org members)* I have manually triggered the GPU
tests with the magic comment `/azp run`.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>