Write particle data in slice plots (#1400)
### Description
Add support for writing particle data in DiagFramePlane. When
`quokka.slice_z.particles` is left empty, default behavior is no
particle types.

This is helpful because in many simulations, slice plot and particle
data is all the information i need to make animation figures.

In the future, we will add support to extract slice plots with deposited
particle fields. However, this won't completely replace the feature
implemented in this PR, because it's not clear how to display particles
in a temperature or velocity field using this approch.


### Related issues
Are there any GitHub issues that are fixed by this pull request? Add a
link to them here.

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues (if applicable; see
above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [x] I have added tests for any new physics that this PR adds to the
code.
- [x] *(For quokka-astro org members)* I have manually triggered the GPU
tests with the magic comment `/azp run`.